Renaming activemessage to active_messages
